GD32: cleanups for casts, includes, and formatting
Add Gd32FirmwareOps.mk with compiler warning flags; switch Includes.mk to use -isystem for external headers and reorder includes. Replace many redundant/static C-style casts with direct negation or proper C++ casts (reinterpret_cast/static_cast) to silence warnings and fix type issues in timers, USB, FATFS and DMX code. Add IRQ/timer defines for GD32F30X_XD and simplify gd32xxxx.h diagnostics and MCU support. Modernize pixel and pixeldmx code (use reinterpret_cast, clearer macros, pointer arithmetic, and consistent bracing/formatting) and minor fixes to USB host MSC error reporting and mount call casting.

+++ lib-gd32/device/usb/f/usb.cpp
@@ -159,16 +159,16 @@ void IntrConfig() {
     // 2 bits for preemption priority, 2 bits for subpriority
     nvic_priority_group_set(NVIC_PRIGROUP_PRE2_SUB2);
     // enable the USB low priority interrupt
-    nvic_irq_enable(static_cast<uint8_t>(USBD_LP_CAN0_RX0_IRQn), 2U, 0U);
+    nvic_irq_enable(USBD_LP_CAN0_RX0_IRQn, 2U, 0U);
     // enable the USB low priority interrupt
-    nvic_irq_enable(static_cast<uint8_t>(USBD_HP_CAN0_TX_IRQn), 1U, 0U);
+    nvic_irq_enable(USBD_HP_CAN0_TX_IRQn, 1U, 0U);
 
     DEBUG_EXIT();
 #else
     DEBUG_ENTRY();
 
     nvic_priority_group_set(NVIC_PRIGROUP_PRE2_SUB2);
-    nvic_irq_enable(static_cast<uint8_t>(USBFS_IRQn), 2U, 0U);
+    nvic_irq_enable(USBFS_IRQn, 2U, 0U);
 
     DEBUG_EXIT();
 #endif
